services/requirements-detection.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
Pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
webcolors==1.13
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
